Module Delegate.For_RPC

The functions in this module are considered too costly to be used in the protocol. They are meant to be used only to answer RPC calls.

val full_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

Returns the full 'balance' of the implicit contract associated to a given key, i.e. the sum of the spendable balance (given by balance or Contract_storage.get_balance) and of the frozen balance of the contract.

The frozen balance is composed of all frozen bonds associated to the contract (given by Contract_storage.get_frozen_bonds), all unstaked frozen deposits, and of the fraction of the frozen deposits that actually belongs to the delegate and not to its stakers.

Only use this function for RPCs: this is expensive.

val delegated_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

val staking_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

val current_cycle_denunciations_list : context -> (public_key_hash * Denunciations_repr.item) list Tezos_protocol_environment_018_Proxford.Lwt.t