Module Tezos_raw_protocol_018_Proxford.Dal_services

val shards : Alpha_context.t -> level:Alpha_context.Raw_level.t -> (Tezos_protocol_environment_018_Proxford.Signature.Public_key_hash.t * (int * int)) list Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

shards ctxt ~level returns the DAL committee as an association list that associates to the public key hash pkh of the member of the committee an interval (s,n), meaning that the slots s;s+1;...;s+n-1 belongs to pkh. It is guaranteed that n>0.