Module Dal_slot_repr.Header

type id = {

published_level : Raw_level_repr.t;

index : Dal_slot_index_repr.t;

}

For Layer-1, a slot is identified by the level at which it is published and the slot's index.

type t = {

id : id;

commitment : Commitment.t;

}

For Layer-1, a slot is described by its slot id and the slot's commitment.

val id_encoding : id Tezos_protocol_environment_018_Proxford.Data_encoding.t

encoding for values of type id.

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

encoding for values of type t.

val pp_id : Tezos_protocol_environment_018_Proxford.Format.formatter -> id -> unit

pretty-printer for values of type id.

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

pretty-printer for values of type t.

val equal : t -> t -> bool

equal function for values of type t.

val verify_commitment : Tezos_protocol_environment_018_Proxford.Dal.t -> Commitment.t -> Commitment_proof.t -> bool Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

verify_commitment cryptobox commitment commitment_proof check that for the given commitment, the commitment proof is correct using the cryptbox primitives.