Module History.Internal_for_tests

val content : t -> Header.t

val proof_statement_is : proof -> [ `Confirmed | `Unconfirmed ] -> bool

proof_statement_is serialized_proof expected will return true if the deserialized proof and the expected proof shape match and false otherwise. Note that it will also return false if deserialization fails.