Module Dal_slot_repr.Page

A DAL slot is decomposed to a successive list of pages with fixed content size. The size is chosen so that it's possible to inject a page in a Tezos L1 operation if needed during the proof phase of a refutation game.

type content = Tezos_protocol_environment_018_Proxford.Bytes.t

type slot_index = Dal_slot_index_repr.t

val pages_per_slot : Tezos_protocol_environment_018_Proxford.Dal.parameters -> int

module Index : sig ... end

val content_encoding : content Tezos_protocol_environment_018_Proxford.Data_encoding.t

Encoding for page contents.

type t = {

slot_id : Header.id;

page_index : Index.t;

}

A page is identified by its slot ID and by its own index in the list of pages of the slot.

type proof = Tezos_protocol_environment_018_Proxford.Dal.page_proof

val equal : t -> t -> bool

equal function for values of type t.

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

encoding for values of type t.

val proof_encoding : proof Tezos_protocol_environment_018_Proxford.Data_encoding.t

encoding for values of type proof.

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

pretty-printer for values of type t.