Module Tezos_raw_protocol_018_Proxford.Delegate_cycles

Per-cycle management of delegates.

val cycle_end : Raw_context.t -> Cycle_repr.t -> (Raw_context.t * Receipt_repr.balance_updates * Tezos_protocol_environment_018_Proxford.Signature.Public_key_hash.t list) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

Trigger the context maintenance at the end of cycle 'n', i.e.: unfreeze the attesting rewards, potentially deactivate delegates. Return the corresponding balances updates and the list of deactivated delegates.

val init_first_cycles : Raw_context.t -> Raw_context.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

init_first_cycles ctxt computes and records the distribution of the total active stake among active delegates. This concerns the total active stake involved in the calculation of baking rights for all cycles in the range 0, preserved_cycles. It also freezes the deposits for all the active delegates.