Module Delegate_sampler.For_RPC

val delegate_baking_power_for_cycle : Raw_context.t -> Cycle_repr.t -> Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> int64 Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

The baking power for a given delegate from the selected stake snapshot of the current cycle.

val delegate_current_baking_power : Raw_context.t -> Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> int64 Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

The baking power for a given delegate computed from its current stake.