Module Tezos_raw_protocol_018_Proxford.Dependent_bool

Dependent booleans

type no = private

| DNo

type yes = private

| DYes

type _ dbool =

| No : no dbool

| Yes : yes dbool

'b dbool is a boolean whose value depends on its type parameter 'b. yes dbool can only be Yes. no dbool can only be No.

type ('a, 'b, 'r) dand =

| NoNo : (no, no, no) dand

| NoYes : (no, yes, no) dand

| YesNo : (yes, no, no) dand

| YesYes : (yes, yes, yes) dand

('a, 'b, 'r) dand is a witness of the logical conjunction of dependent booleans. 'r is the result of 'a and 'b.

type ('a, 'b) ex_dand =

| Ex_dand : ('a, 'b, _) dand -> ('a, 'b) ex_dand

val dand : 'a dbool -> 'b dbool -> ('a, 'b) ex_dand

Logical conjunction of dependent booleans.

val dbool_of_dand : ('a, 'b, 'r) dand -> 'r dbool

Result of the logical conjunction of dependent booleans.

type (_, _) eq =

| Eq : ('a, 'a) eq

Type equality witness.

val merge_dand : ('a, 'b, 'c1) dand -> ('a, 'b, 'c2) dand -> ('c1, 'c2) eq

merge_dand proves that the type dand represents a function, i.e. that there is a unique 'r such that ('a, 'b, 'r) dand is inhabited for a given 'a and a given 'b.