Module Tezos_raw_protocol_018_Proxford.Entrypoint_repr

type t

An entrypoint is a non-empty string of at most 31 characters

val compare : t -> t -> int

Total ordering of entrypoints

val (=) : t -> t -> bool

Equality of entrypoints

val default : t

Default entrypoint "default"

val is_default : t -> bool

Checks whether an entrypoint is the default entrypoint

val root : t

Root entrypoint "root"

val is_root : t -> bool

Checks whether an entrypoint is the root entrypoint

val do_ : t

Entrypoint "do"

val set_delegate : t

Entrypoint "set_delegate"

val remove_delegate : t

Entrypoint "remove_delegate"

val deposit : t

Deposit entrypoint "deposit"

val stake : t

Entrypoint "stake"

val unstake : t

Entrypoint "unstake"

val finalize_unstake : t

Entrypoint "finalize_unstake"

val set_delegate_parameters : t

Entrypoint "set_delegate_parameters"

val is_deposit : t -> bool

Checks whether an entrypoint is the deposit entrypoint

val of_annot_strict : loc:Script_repr.location -> Non_empty_string.t -> t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

Converts an annot to an entrypoint. Returns an error if the string is too long or is "default".

val of_string_strict : loc:Script_repr.location -> string -> t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

Converts a string to an entrypoint. Returns an error if the string is too long or is "default". Converts "" to "default".

val of_string_strict_exn : string -> t

Converts a string to an entrypoint. Fails with Invalid_arg if the string is too long or is "default". Converts "" to "default".

val of_annot_lax : Non_empty_string.t -> t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

Converts an annot to an entrypoint. Returns an error if the string is too long. Accepts "default".

val of_annot_lax_opt : Non_empty_string.t -> t option

Converts an annot to an entrypoint. Returns None if the string is too long. Accepts "default".

val of_string_lax : string -> t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

Converts a string to an entrypoint. Returns an error if the string is too long. Accepts "default" and converts "" to "default".

val to_non_empty_string : t -> Non_empty_string.t

Converts an entrypoint to a non-empty string. "default" is kept as is.

val to_string : t -> string

Converts an entrypoint to a string. "default" is kept as is.

val to_address_suffix : t -> string

Converts an entrypoint to a string used as an address suffix. For the default entrypoint, the result is the empty string. Otherwise it is "%" followed by the entrypoint.

val unparse_as_field_annot : t -> string

Converts an entrypoint to a string used as a field annotation of a parameter union type. It is "%" followed by the entrypoint. The default entrypoint is converted to "%default".

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

Pretty-print an entrypoint

val simple_encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

An encoding of entrypoints reusing the lax semantics. Decoding fails if the string is too long. "" is decoded into "default". "default" is encoded into "default".

val value_encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

An encoding of entrypoints reusing the strict semantics. Decoding fails if the string is too long or is "default". "" is decoded into "default". "default" is encoded into "".

val smart_encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

An optimized encoding of entrypoints, used for operations.

val rpc_arg : t Tezos_protocol_environment_018_Proxford.RPC_arg.t

Entrypoint RPC arg.

val in_memory_size : t -> Saturation_repr.may_saturate Saturation_repr.t

In-memory size of an entrypoint

module Set : Tezos_protocol_environment_018_Proxford.Set.S with type elt = t

Set of entrypoints

module Map : Tezos_protocol_environment_018_Proxford.Map.S with type key = t

Map of entrypoints