Module Tezos_raw_protocol_018_Proxford.Frozen_staker_repr

type t = private

| Baker of Tezos_protocol_environment_018_Proxford.Signature.public_key_hash
(*
The baker itself.
*)

| Single_staker of {

staker : Contract_repr.t;

delegate : Tezos_protocol_environment_018_Proxford.Signature.public_key_hash;

}
(*
A single staker, cannot be the delegate.
*)

| Shared_between_stakers of {

delegate : Tezos_protocol_environment_018_Proxford.Signature.public_key_hash;

}
(*
The delegate's stakers simultaneously (delegate excluded).
*)

Adding and removing stake can be done from/toward a delegate, one of its staker, or all its stakers at once. We need to distinguish these cases to enforce the staking over baking limit.

val baker : Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> t

val single_staker : staker:Contract_repr.t -> delegate:Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> t

val shared_between_stakers : delegate:Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> t

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

val compare : t -> t -> int