Module Path_encoding.Make_hex

Path encoding in hex: /[0-9a-f]{2}+/

Parameters

module H : ENCODING

Signature

val to_path : H.t -> string list -> string list

to_path t postfix returns the context path name for t postfixed with postfix

val of_path : string list -> H.t option

of_path path parses path as a context path name for t

val path_length : int

Directory levels of the path encoding of t