Module Tezos_raw_protocol_018_Proxford.Raw_context_intf

All context manipulation functions. This signature is included as-is for direct context accesses, and used in Storage_functors to provide restricted views to the context.

type depth = [

| `Eq of int

| `Le of int

| `Lt of int

| `Ge of int

| `Gt of int

]

The tree depth of a fold. See the fold function for more information.

type config = Tezos_protocol_environment_018_Proxford.Context.config

The type for context configuration. If two trees or stores have the same configuration, they will generate the same context hash.

module type VIEW = sig ... end

module Kind : sig ... end

module type TREE = sig ... end

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

module type PROOF = sig ... end

Proofs are compact representations of trees which can be shared between peers.

module type T = sig ... end