Module Sc_rollup_commitment_storage.Internal_for_tests

val get_cemented_commitments_with_levels : Raw_context.t -> Sc_rollup_repr.t -> ((Commitment_hash.t * Raw_level_repr.t) list * Raw_context.t) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

get_cemented_commitments_with_levels ctxt rollup returns a list of all cemented commitment hashes and corresponding inbox levels that are present in the storage, ordered by inbox level.

May fail with:

Sc_rollup_does_not_exist if rollup does not exist