Module Tezos_raw_protocol_018_Proxford.Sc_rollup_inbox_merkelized_payload_hashes_repr

type Tezos_protocol_environment_018_Proxford.Error_monad.error +=

| Merkelized_payload_hashes_proof_error of string

module Hash : Tezos_protocol_environment_018_Proxford.S.HASH with type t = Tezos_protocol_environment_018_Proxford.Smart_rollup.Merkelized_payload_hashes_hash.t

type t

A type representing the head of a merkelized list of Sc_rollup_inbox_message_repr.serialized message. It contains the hash of the payload and the index on the list.

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

type merkelized_and_payload = {

merkelized : t;

payload : Sc_rollup_inbox_message_repr.serialized;

}

module History : sig ... end

A History.t is a lookup table of merkelized_and_payloads. Payloads are indexed by their hash Hash.t. This history is needed in order to produce proof.

val hash : t -> Hash.t

hash merkelized is the hash of merkelized. It is used as key to remember a merkelized payload hash in an History.t.

val remember : History.t -> t -> Sc_rollup_inbox_message_repr.serialized -> History.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

remember history merkelized payload remembers the {merkelized; payload} in history with key hash merkelized.

val genesis_no_history : Sc_rollup_inbox_message_repr.serialized -> t

genesis_no_history payload is the initial merkelized payload hashes with index 0.

val genesis : History.t -> Sc_rollup_inbox_message_repr.serialized -> (History.t * t) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

genesis history payload is the initial merkelized payload hashes with index 0. It is remembered in history using remember.

val add_payload_no_history : t -> Sc_rollup_inbox_message_repr.serialized -> t

add_payload_no_history merkelized payload creates a new t with payload and merkelized as ancestor (i.e. index = succ (get_index merkelized)).

val add_payload : History.t -> t -> Sc_rollup_inbox_message_repr.serialized -> (History.t * t) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

add_payload is identical to add_payload_no_history but the resulting merkelized is remembered in history with remember.

val equal : t -> t -> bool

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

val get_payload_hash : t -> Sc_rollup_inbox_message_repr.Hash.t

get_payload_hash merkelized returns the Sc_rollup_inbox_message_repr.serialized payload's hash of merkelized.

val get_index : t -> Tezos_protocol_environment_018_Proxford.Z.t

get_index merkelized returns the index of merkelized.

type proof = private t list

Given two t (a, b) and a Sc_rollup_inbox_message_repr.serialized payload, a proof guarantees that payload hash is equal to a and that a is an ancestor of b; i.e. get_index a < get_index b.

val pp_proof : Tezos_protocol_environment_018_Proxford.Format.formatter -> proof -> unit

val proof_encoding : proof Tezos_protocol_environment_018_Proxford.Data_encoding.t

val produce_proof : History.t -> index:Tezos_protocol_environment_018_Proxford.Z.t -> t -> (merkelized_and_payload * proof) option

produce_proof history ~index into_ returns a merkelized_and_payload with index index and a proof that it is an ancestor of into_. Returns None if no merkelized payload with index is found (either in the history or index is not inferior to get_index into_).

val verify_proof : proof -> (t * t) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

verify_proof proof returns (a, b) where proof validates that a is an ancestor of b. Fails when proof is not a valid inclusion proof.

module Internal_for_tests : sig ... end