Module Tezos_raw_protocol_018_Proxford.Sc_rollup_repr

The basic components of an optimistic rollup for smart-contracts.

An optimistic rollup for smart-contracts is made of two main components:

a proof generating virtual machine (PVM), which provides the essential semantics for the rollup operations to be validated by the layer 1 in case of dispute about a commitment ;

a database which maintains the cemented operations of the rollup as well as the potentially-disputed operations.

module Address : sig ... end

A smart rollup has an address starting with "sr1".

module Internal_for_tests : sig ... end

module State_hash : module type of struct include Tezos_protocol_environment_018_Proxford.Smart_rollup.State_hash end

module Number_of_ticks : sig ... end

Number of ticks computed by a single commitment. This represents a claim about the state of the PVM, which can be disputed as part of a commitment dispute.

type t = Address.t

A smart contract rollup is identified by its address.

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

val rpc_arg : t Tezos_protocol_environment_018_Proxford.RPC_arg.t

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

val in_memory_size : t -> Cache_memory_helpers.sint

in_memory_size sc_rollup returns the number of bytes sc_rollup uses in RAM.

module Staker : sig ... end

A Staker is an implicit account, identified by its public key hash.

module Index : Storage_description.INDEX with type t = Address.t

The data model uses an index of these addresses.