Module Tezos_raw_protocol_018_Proxford.Sc_rollup_whitelist_repr

type t = Tezos_protocol_environment_018_Proxford.Signature.Public_key_hash.t list

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

val pp : Tezos_protocol_environment_018_Proxford.Format.formatter -> t -> unit

type last_whitelist_update = {

message_index : Tezos_protocol_environment_018_Proxford.Z.t;

outbox_level : Raw_level_repr.t;

}

This type is used in the `Storage.Last_whitelist_update. It is the latest whitelist update message executed in the L1. It is used to prevent user to execute whitelist update in unchronological order.

val last_whitelist_update_encoding : last_whitelist_update Tezos_protocol_environment_018_Proxford.Data_encoding.t

type update =

| Public

| Private of t

val update_encoding : update Tezos_protocol_environment_018_Proxford.Data_encoding.t