Module Tezos_raw_protocol_018_Proxford.Sc_rollups

Here is the list of PVMs available in this protocol.

module PVM : sig ... end

module Kind : sig ... end

A smart contract rollup has a kind, which assigns meaning to rollup operations.

val genesis_state_hash_of : boot_sector:string -> Kind.t -> Sc_rollup_repr.State_hash.t Tezos_protocol_environment_018_Proxford.Lwt.t

genesis_hash_of machine ~boot_sector computes the initial state hash of a rollup given an initial boot_sector.