Module Data_unparser.Internal_for_benchmarking

For benchmarking purpose, we also export versions of the unparsing functions which don't call location stripping. These functions are not carbonated and should not be called directly from the protocol.

val unparse_data : Alpha_context.context -> stack_depth:int -> unparsing_mode -> ('a, 'ac) Script_typed_ir.ty -> 'a -> (Alpha_context.Script.node * Alpha_context.context) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

val unparse_code : Alpha_context.context -> stack_depth:int -> unparsing_mode -> Alpha_context.Script.node -> (Alpha_context.Script.node * Alpha_context.context) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t