Module Tezos_raw_protocol_018_Proxford.Stake_repr

type t = private {

frozen : Tez_repr.t;

weighted_delegated : Tez_repr.t;

}

Stake of a delegate.

val zero : t

val make : frozen:Tez_repr.t -> weighted_delegated:Tez_repr.t -> t

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

val get_frozen : t -> Tez_repr.t

Returns only the frozen part of a stake

val (+?) : t -> t -> t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

val staking_weight : t -> int64

The weight of a staker or a set of stakers. Since this function is applied on a Stake_repr.t, the limits should already have been applied using apply_limits if necessary.

val compare : t -> t -> int