Module Token.Internal_for_tests

val allocated : Raw_context.t -> container -> (Raw_context.t * bool) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

allocated ctxt container returns a new context because of possible access to carbonated data, and a boolean that is true when balance ctxt container is guaranteed not to fail, and false when balance ctxt container may fail.

type container_with_balance = [

| `Contract of Contract_repr.t

| `Collected_commitments of Blinded_public_key_hash.t

| `Block_fees

| `Frozen_bonds of Contract_repr.t * Bond_id_repr.t

]

val balance : Raw_context.t -> [< container_with_balance ] -> (Raw_context.t * Tez_repr.t) Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

balance ctxt container returns a new context because of an access to carbonated data, and the balance associated to the token holder. This function may fail if allocated ctxt container returns false. Returns an error with the message "get_balance" if container refers to an originated contract that is not allocated.