Module Tezos_raw_protocol_018_Proxford.Unstaked_frozen_deposits_repr

Datatype for a map from cycle to deposits, where all unslashable cycles are squashed.

Expected to be used for a small number of cycles at a time, typically bounded by preserved_cycles + max_slashing_period plus a small constant.

type t

Storable version.

type squashed = private {

unslashable_cycle : Cycle_repr.t option;

t : t;

}

To be used locally, do not preserve values of this type over cycles.

val empty : unslashable_cycle:Cycle_repr.t option -> squashed

val encoding : t Tezos_protocol_environment_018_Proxford.Data_encoding.t

val squash_unslashable : unslashable_cycle:Cycle_repr.t option -> t -> squashed Tezos_protocol_environment_018_Proxford.Error_monad.tzresult

Once read, t must be converted to squashed with squash_unslashable to be used efficiently. For a given unslashable_cycle, squash_unslashable ~unslashable_cycle is idempotent.

val get : Cycle_repr.t -> squashed -> Deposits_repr.t

val update : f: (Deposits_repr.t -> Deposits_repr.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult) -> Cycle_repr.t -> squashed -> squashed Tezos_protocol_environment_018_Proxford.Error_monad.tzresult