Module Tezos_raw_protocol_018_Proxford.Unstaked_frozen_deposits_storage

Simple abstraction from low-level storage to handle unstaked frozen deposits.

This module is responsible for maintaining the Storage.Contract.Unstaked_frozen_deposits table.

val balance : Raw_context.t -> Tezos_protocol_environment_018_Proxford.Signature.Public_key_hash.t -> Cycle_repr.t -> Tez_repr.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

balance ctxt delegate cycle returns the amount of unstaked frozen deposits for delegate at cycle. If cycle is an unslashable cycle, the returned amount is the squashed amount of all the unslashable cycles.

val get : Raw_context.t -> Tezos_protocol_environment_018_Proxford.Signature.Public_key_hash.t -> Cycle_repr.t -> Deposits_repr.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

get acts like balance but returns both the initial amount and the current amount.

val credit_only_call_from_token : Raw_context.t -> Unstaked_frozen_staker_repr.t -> Cycle_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

credit_only_call_from_token ctxt staker cycle amount credits the unstaked frozen deposits for staker at cycle by amount. If cycle is an unslashable cycle, the credited cycle is the last unslashable cycle.

val spend_only_call_from_token : Raw_context.t -> Unstaked_frozen_staker_repr.t -> Cycle_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

spend_only_call_from_token ctxt staker cycle amount spends amount from the unstaked frozen deposits for staker at cycle. If cycle is an unslashable cycle, the amount is spent from the last unslashable cycle. The function returns the error Subtraction_underflow if the balance is too low.

val decrease_initial_amount_only_for_stake_from_unstake : Raw_context.t -> Tezos_protocol_environment_018_Proxford.Signature.public_key_hash -> Cycle_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_018_Proxford.Error_monad.tzresult Tezos_protocol_environment_018_Proxford.Lwt.t

decrease_initial_amount_only_for_stake_from_unstake ctxt staker cycle amount decreases amount from the unstaked frozen deposits for staker at cycle. It is only called if the cycle hasn't been slashed, so the amount removed is the same as the amount spent in current_amount.